mach-exynos/Kconfig: Force cpuidle to be disabled on HKDK4412
We are enforing this policy to prevent the infamous reboot bug on cpuidle.
Until this bug is fixed we can safely disable CPU_IDLE for ODROIDs.

The cost of leaving CPUIDLE enabled also doesn't make sense on a dev board that is always on.
The amount of battery saved by this will be marginal and people using ODROID's on battery probably won't notice the change.

This will also create a better user experience