odroidu2-fan: This patch fixes the ODROID-U2 PWM Fan
Removed the mutex from the fan driver, since the pwm should take care of this.

Enabled Timer0 on platform file.

samsung_bl_data should be called later to not mess the creation order of the timers.

duty_ns can't be 0. If we get a duty of 0 it should be setted to at least 1.
That's how the Linux PWM api handles it.

Minimal logic fix on fan driver